+++ clang-tools-extra/test/clang-tidy/readability-redundant-pointer-dereference-chain.cpp
@@ -1,0 +1,20 @@
+// RUN: %check_clang_tidy %s readability-redundant-pointer-dereference-chain %t
+
+struct T {
+  int i;
+  T *tp;
+};
+
+T *talloc();
+
+void dummy() {
+  T *t1 = talloc();
+  if (!t1)
+    return;
+  T *t1n = t1->tp;
+  if (!t1n)
+    return;
+  int i = t1n->i;
+  // CHECK-MESSAGES: :[[@LINE-7]]:6: warning: foo bar baz qux [readability-redundant-pointer-dereference-chain]
+  // CHECK-MESSAGES: :[[@LINE-5]]:6: warning: foo bar baz qux [readability-redundant-pointer-dereference-chain]
+}